pyproject.toml
[project]
name = "zen-study-scraper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
